Refactor RPC to support seperate protocols
