+++ node/forest_libp2p/src/blocksync/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2020 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+mod message;
+
+pub use self::message::*;
+
+pub const BLOCKSYNC_PROTOCOL_ID: &[u8] = b"/fil/sync/blk/0.0.1";
