+++ node/forest_libp2p/src/hello/mod.rs
@@ -1,0 +1,4 @@
+// Copyright 2020 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+pub const HELLO_PROTOCOL_ID: &[u8] = b"/fil/hello/1.0.0";
