+++ node/forest_libp2p/src/lib.rs
@@ -4,7 +4,9 @@
 #![recursion_limit = "1024"]
 
 mod behaviour;
+pub mod blocksync;
 mod config;
+pub mod hello;
 pub mod rpc;
 mod service;
