+++ node/forest_libp2p/src/rpc/mod.rs
@@ -2,14 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 mod behaviour;
-mod blocksync_message;
 mod codec;
 mod handler;
 mod protocol;
 mod rpc_message;
 
 pub use behaviour::*;
-pub use blocksync_message::*;
 pub use codec::*;
 pub use handler::*;
 pub use protocol::*;
